requirements.txt
python-telegram-bot[all]==21.10
elevenlabs==1.50.5
httpx==0.28.1
pydantic-settings==2.8.1
pytest==8.3.5
pytest-asyncio==0.26.0
